feat(similarityIndex): add new getSimilarityToLibs()
This function produces four types of similarities between unknown signature and
all libraries+versions in our database of libraries.

The four types of similarities are:
  * our metric on set of fn names
  * jaccard metric on set of fn names
  * fn comparison metric by fn statement tokens
  * fn comparison metric by fn statement types

Right now, librarySimilarityByFunctionStatementTokens() seems to have
a bug in it, and librarySimilarityByFunctionStatementTypes() is not
implemented yet, but will be the same as by fn statement tokens.