refactor(similarityIndex:fnStToks): mark matched function with __matched
Before, the matched function was being removed from list of all functions, so
that it is not assigned second time. However, this way, indexes were shifting,
and it was hard to construct mapping between library functions and unknown
functions. Now, we are marking whichever function is matched with __matched
property.