fix(bin/find-lib-mentions.worker): create and destroy RegExp on demand
Before, all RegExp objects were created in the beginning of the script, because
they were not supposed to change. However, they change and cause memory in use
to increase a lot, so that workers run out of heap memory. So, instead of using
same instances of RegExp, new instance of RegExp is created for each for-loop.
This slows down the entire script, as it now creates, deletes and re-creates
a lot of RegExp objects. However, it seems that workers do not run out of
memory anymore.