feat(bundle_similarity): create secondary matches with <100% sim
In the first stage of matching, we are trying to find all libraries from
candidate list which have 100% matching. If some candidate library does not
have 100% match as top match, then this library is moved into second stage. In
second stage, libraries are allowed to have not 100% match, and they are still
saved in the results.