refactor(fnStToks:v6:perf): mark unknown els as matched
Instead of indexing every element and removing it from array when it got
matched, we are now marking all elements as unmatched, then those that got
matched are being marked as matched and skipped in later matching procedures.
This improves the performance greatly, because we know which element needs to
be marked as matched, and it can be done in O(1) time, while filtering would
take O(n).