refactor(fnStToks:v6:perf): build possible indexes along with matching
This way, we do not need to iterate over unknown array again in order to
determine which values got matched and which did not get matched. This method
initializes array of possible indexes of the same size as unknown array with
'-1', which corresponds to unmatched marker, and after each individual function
is matched, it updates array of possible indexes at given index from unknown
array with matched index from lib array in O(1) time.