feat(parseApps/cordova): worker loads signature and candidates
If we are given pool, which implements proper method, the worker can read
corresponding signature and candidate files. So, signature file path and
candidate file path are passed. Also, if worker returns 'true' value, it means
that there were no candidates in the candidate file, and master function will
report that.